--- src/About.js
@@ -1,17 +1,0 @@
-import React from 'react';
-
-class About extends React.Component {
-  constructor(props) {
-    super(props);
-  }
-
-  render() {
-    return (
-      <>
-       
-      </>
-    );
-  }
-}
-
-export default About;
