+++ src/About/About.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class About extends React.Component {
+    constructor(props) {
+        super(props);
+    }
+
+    render() {
+        return <></>;
+    }
+}
+
+export default About;
